refactor: implement checkout, fulfillment, and payment controllers with order management functionality
